Fix infinite loop in NamedPattern on Android (REGEX-9)
Matcher.reset() in Android does not reload the string input and continues using
the original value. A loop in buildStandardPattern() checks whether there are
any remaining named capture groups that need to be converted into unnamed
capture groups. On each iteration, it calls Matcher.reset() (without arguments)
to reload the newly updated string in the matcher, but the reload was not
occurring in Android, thus the Matcher was finding the search target in the
original string and repeating itself.

This change passes the updated string as an argument to Matcher.reset(String),
which correctly updates the matcher's string input (the same as the desktop
version of it).